Fix Figshare upload progress bar
Add `total` argument to `Progress.progress`, and make the figshare progress update include that. Previously it was doing the fraction of total upload completed, but the task status formatter was casting it to an `int`, so it was always `0`.

Also, update the React components to use `total`, and move the progress tracking table into its own component.