Replace Figshare OAuth with personal tokens
Remove the FigshareAuthorization model and replace it with the
figshare_personal_token field on the User model. The other fields on
the FigshareAuthorization model are available through the Figshare API
(figshare_account_id) or are not needed anymore. Update the UI to
replace the OAuth button and page with a modal to add a personal access
token.

The Figshare OAuth flow inexplicably stopped working, so we are
replacing OAuth tokens with personal tokens. This also solves the
problem we were having where connecting one Taiga account to our shared
Figshare account would invalidate any other Taiga accounts' tokens
linked to that Figshare account.